pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diffusion-core"
version = "0.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.0",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pre-commit>=3.7",
  "pytest>=8.0",
  "pyright>=1.1",
  "ruff>=0.5",
  "mkdocs>=1.5",
  "mkdocs-material>=9",
]
docs = [
  "mkdocs>=1.5",
  "mkdocs-material>=9",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["diffusion_core*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
markers = [
  "slow: long-running tests (excluded from fast suite)",
  "integration: hits filesystem/network/long loops (excluded from fast suite)",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]
force-exclude = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []
per-file-ignores = { "src/diffusion_core/__init__.py" = ["F401"] }

[tool.ruff.lint.pydocstyle]
convention = "google"